Split `stance_extraction.py` into 3 smaller files
The single file was too large to work with easily. I have split out each
class into its own file.

I left out the `__author__` file header because this project is the work
of multiple people.
